Remove redundent comments

+++ contracts/core/LiquidityMining/GaugeController/PendleGaugeControllerSidechain.sol
@@ -14,9 +14,9 @@ contract PendleGaugeControllerSidechain is PendleGaugeController, CelerReceiver 
     ) PendleGaugeController(_pendle, _marketFactory) CelerReceiver(_governanceManager) {}
 
     function _executeMessage(bytes memory message) internal virtual override {
-        (uint128 timestamp, address[] memory markets, uint128[] memory incentives) = abi.decode(
+        (uint128 timestamp, address[] memory markets, uint256[] memory incentives) = abi.decode(
             message,
-            (uint128, address[], uint128[])
+            (uint128, address[], uint256[])
         );
         _receiveVotingResults(timestamp, markets, incentives);
     }
